switch to c++20 and some cleanup

+++ Source/egt.cpp
@@ -42,7 +42,7 @@ bool initTB() {
    Logging::LogIt(Logging::logInfo) << "Init tb from path " << DynamicConfig::syzygyPath;
    if (!tb_init(DynamicConfig::syzygyPath.c_str())) return MAX_TB_MEN = 0, false;
    else
-      MAX_TB_MEN = TB_LARGEST;
+      MAX_TB_MEN = static_cast<int>(TB_LARGEST);
    Logging::LogIt(Logging::logInfo) << "MAX_TB_MEN: " << MAX_TB_MEN;
    return true;
 }
